Modifications to get the Mini Arm working
MiniArm.ino is working, controls the lower 3 motors. MiniGripper is not working, needs to be ported to work similarly to MiniArm and control the upper 3 motors. Arm/Gripper.ino are not working and need to be ported to use the new changes, primarily, the difference in the constructor.

See MiniArm.ino for a list of TODOs.